requirements.txt
asgiref==3.8.1
Django==3.1.14
pillow==11.1.0
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
